requirements.txt
Django==5.0.2
pip==24.0
attrs==23.2.0
wheel==0.42.0
pillow==10.2.0
filelock==3.13.1
asgiref==3.7.2
sqlparse==0.4.4
psycopg2==2.9.9
colorama==0.4.6
numpy==1.26.4
requests==2.22.0
typing_extensions==4.12.2
Markdown==3.7
djangorestframework==3.15.2
django-filter==24.3
pytz==2024.1
django-debug-toolbar==4.4.6
cffi==1.16.0
future==0.18.3
setuptools==69.1.0
certifi==2021.10.8
chardet==3.0.4
urllib3==1.25.11
idna==2.8
six==1.16.0
PyJWT==1.7.1
Click==7.0
python-dateutil==2.8.2
pycparser==2.22
packaging==23.2
platformdirs==4.2.0
django-appconf==1.0.6